Support make DESTDIR=... install for easier packaging
Packaging scripts often use the DESTDIR=/some/directory variable to make the
build system install the binaries in a specific directory which will be
compressed to create the package. This commit adds support for this variable,
by introducing a new INSTALL_DIR variable in the Makefile and replacing PREFIX
by INSTALL_DIR everywhere it makes sense.

Note that by default, INSTALL_DIR=$(PREFIX) so there should not be any
compatibility problem with this commit.