[fix] session,memleak,opa2js: client context was not propagated resulting in server context
The memory leak concerns non top-level sessions, which following a GET request on xhtml content (i.e. a web page), appeared in the xhtml serialization.
The situation also implicitly happened when having a non serializable server closure is included in the xhtml (via onclick ...).

In that case, the sessions were not collected when the client quit the web page, and would remains in memory in the server.