[enhance] iconv: Change iconv interface for errors.
If iconv encountered an error while converting a string, the code
would throw an exception, while now it should return none. Thus, the
return type now is option(binary) instead of just binary.

There is, however, a difference right now in the behavior of the two
backends. In ocaml, converting "ça va" from utf-8 to ascii yields "??a
va", while in the node backend it returns "none". We should make both
interfaces converge in the future.