[fix] session: lexical scope problem.
Blocks in JS don't bind local variables, thus we need to wrap all new
scopes in "(function () { })();" blocks.

There is a weird interaction between global variables declared inside
blocks and the AST cleaning, which was erasing the definition of some
variables. Fixing the scope seems to make this problem go away for now
in this file.