[fix] mail: give required `-t` argument to sendmail
Previous versions of nodemailer.js used to add `-t` automatically to
sendmail command line, in order to fetch recipients from the `To:`,
`Cc:` and `Bcc:` fields. However, this has been recently removed so that
recipients can be specified on the command line instead. It is now
necessary for the caller to add `-t`, or sendmail will exit with error
75 because of the lack of recipients.

Compatibility with older sendmail versions are preserved as giving
multiple `-t` on the command line is similar to giving only one.