[enhance] facebook: allow the use of a dispatcher
The Facebook token retriever function requires a raw unparsed query
string that it parses immediately. This prevents users from using a
dispatcher because the query string has been parsed already.

This change allows either a query string or a list of tuples to be used
when handling the answer coming from Facebook redirection.