[enhance] xml_parser: allow trx-parser-based attribute patterns to bind variables
E.g. <a x=(y=some_parser) />: print(y)
would fail with "y is unbound", and now works

This is however limited to single patterns.
<a x=(y=some_parser)|(y=some_parser) />
will still give "y is unbound".

/!\ Since attributes are bound before matching its value, the following patterns are equivalent:
<a x="{x}" />
<a x=_ />
Indeed, "{x}" refers to the value of the attribute x.
Todo: we should produce a warning in such a case.